Cause FETCH 1 to return the current cursor row, or zero if at beginning/end of cursor.
Have MOVE return 0/1 depending on cursor position.

Matches SQL spec.

Pass cursor counter from parser as a long rather than int.

Doc updates.

+++ src/include/commands/portalcmds.h
@@ -25,7 +25,7 @@
  *		BadArg if forward invalid.
  *		"ERROR" if portal not found.
  */
-extern void PerformPortalFetch(char *name, bool forward, int count,
+extern void PerformPortalFetch(char *name, bool forward, long count,
 				   CommandDest dest, char *completionTag);
 
 /*
